[Engine] Moved IndexingDictionary to Xenko.Core
