+++ sources/core/Xenko.Core/Serialization/Serializers/IndexingDictionarySerializer.cs
@@ -1,8 +1,9 @@
-﻿// Copyright (c) Xenko contributors (https://xenko.com) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
+// Copyright (c) Xenko contributors (https://xenko.com) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
-using Xenko.Core.Serialization.Serializers;
 
-namespace Xenko.Engine
+using Xenko.Core.Collections;
+
+namespace Xenko.Core.Serialization.Serializers
 {
     internal class IndexingDictionarySerializer<TValue> : DictionaryAllSerializer<IndexingDictionary<TValue>, int, TValue> where TValue : class
     {
