+++ sources/engine/Xenko.Engine/Rendering/IModelInstance.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Xenko contributors (https://xenko.com) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
 
-using Xenko.Engine;
+using Xenko.Core.Collections;
 
 namespace Xenko.Rendering
 {
